demos/polls/requirements.txt
-e .
pytest-aiohttp==0.1.3
flake8
marshmallow
